requirements.txt

streamlit
pandas
openpyxl
scikit-learn
matplotlib
python-dateutil
pytest
pre-commit
black
isort
flake8
